serde crate compiles cargo test
