+++ tiledb/api/src/lib.rs
@@ -76,7 +76,7 @@ pub use context::{Context, ContextBound};
 pub use datatype::Datatype;
 pub type Result<T> = std::result::Result<T, error::Error>;
 
-#[cfg(feature = "serde")]
+#[cfg(any(test, feature = "serde"))]
 pub trait Factory {
     type Item;
