+++ tiledb/api/src/tests/mod.rs
@@ -1,19 +1,21 @@
 pub mod examples;
 
 pub mod prelude {
-    pub use crate::array::attribute::{
-        AttributeData, Builder as AttributeBuilder,
+    pub use tiledb_common::array::dimension::DimensionConstraints;
+    pub use tiledb_common::array::{
+        ArrayType, CellOrder, CellValNum, Mode, TileOrder,
     };
-    pub use crate::array::dimension::{
-        Builder as DimensionBuilder, DimensionConstraints, DimensionData,
-    };
-    pub use crate::array::domain::{Builder as DomainBuilder, DomainData};
-    pub use crate::array::schema::{Builder as SchemaBuilder, SchemaData};
-    pub use crate::array::{
-        Array, ArrayType, Attribute, CellOrder, CellValNum, Dimension, Domain,
-        Mode, Schema, TileOrder,
-    };
-    pub use crate::Datatype;
+    pub use tiledb_common::datatype::Datatype;
+    pub use tiledb_serde::array::attribute::AttributeData;
+    pub use tiledb_serde::array::dimension::DimensionData;
+    pub use tiledb_serde::array::domain::DomainData;
+    pub use tiledb_serde::array::schema::SchemaData;
+
+    pub use crate::array::attribute::Builder as AttributeBuilder;
+    pub use crate::array::dimension::Builder as DimensionBuilder;
+    pub use crate::array::domain::Builder as DomainBuilder;
+    pub use crate::array::schema::Builder as SchemaBuilder;
+    pub use crate::array::{Array, Attribute, Dimension, Domain, Schema};
 
     pub use crate::query::{
         Query, QueryBuilder, QueryLayout, ReadBuilder, ReadQuery, WriteBuilder,
