+++ tiledb/serde/Cargo.toml
@@ -17,10 +17,14 @@ tiledb-test-utils = { workspace = true, optional = true }
 tiledb-sys = { workspace = true }
 
 [dev-dependencies]
+num-traits = { workspace = true }
 proptest = { workspace = true }
+tiledb-common = { workspace = true, features = ["proptest-strategies"] }
+tiledb-test-utils = { workspace = true }
+tiledb-utils = { workspace = true }
 
 [features]
 default = []
 option-subset = ["dep:tiledb-proc-macro", "dep:tiledb-utils", "tiledb-common/option-subset"]
-proptest-strategies = ["dep:num-traits", "dep:proptest", "dep:tiledb-test-utils"]
+proptest-strategies = ["dep:num-traits", "dep:proptest", "dep:tiledb-test-utils", "dep:tiledb-utils", "tiledb-common/proptest-strategies"]
 serde = ["dep:serde", "tiledb-common/serde"]
