+++ tiledb/serde/src/array/schema/mod.rs
@@ -181,75 +181,3 @@ impl std::iter::FusedIterator for FieldDataIter<'_> {}
 
 #[cfg(any(test, feature = "proptest-strategies"))]
 pub mod strategy;
-
-#[cfg(test)]
-mod tests {
-    /// Test what the default values filled in for `None` with schema data are.
-    /// Mostly because if we write code which does need the default, we're expecting
-    /// to match core and need to be notified if something changes or we did something
-    /// wrong.
-    #[test]
-    fn test_defaults() {
-        use crate::array::dimension::DimensionConstraints;
-
-        let ctx = Context::new().unwrap();
-
-        let dense_spec = SchemaData {
-            array_type: ArrayType::Dense,
-            domain: DomainData {
-                dimension: vec![DimensionData {
-                    name: "d".to_string(),
-                    datatype: Datatype::Int32,
-                    constraints: DimensionConstraints::Int32([0, 100], None),
-                    filters: None,
-                }],
-            },
-            attributes: vec![AttributeData {
-                name: "a".to_string(),
-                datatype: Datatype::Int32,
-                ..Default::default()
-            }],
-            ..Default::default()
-        };
-
-        let dense_schema = dense_spec
-            .create(&ctx)
-            .expect("Error creating schema from mostly-default settings");
-
-        assert_eq!(ArrayType::Dense, dense_schema.array_type().unwrap());
-        assert_eq!(10000, dense_schema.capacity().unwrap());
-        assert_eq!(CellOrder::RowMajor, dense_schema.cell_order().unwrap());
-        assert_eq!(TileOrder::RowMajor, dense_schema.tile_order().unwrap());
-        assert!(!dense_schema.allows_duplicates().unwrap());
-
-        let sparse_spec = SchemaData {
-            array_type: ArrayType::Sparse,
-            domain: DomainData {
-                dimension: vec![DimensionData {
-                    name: "d".to_string(),
-                    datatype: Datatype::Int32,
-                    constraints: DimensionConstraints::Int32([0, 100], None),
-                    filters: None,
-                }],
-            },
-            attributes: vec![AttributeData {
-                name: "a".to_string(),
-                datatype: Datatype::Int32,
-                ..Default::default()
-            }],
-            ..Default::default()
-        };
-        let sparse_schema = sparse_spec
-            .create(&ctx)
-            .expect("Error creating schema from mostly-default settings");
-
-        assert_eq!(ArrayType::Sparse, sparse_schema.array_type().unwrap());
-        assert_eq!(
-            SchemaData::DEFAULT_SPARSE_TILE_CAPACITY,
-            sparse_schema.capacity().unwrap()
-        );
-        assert_eq!(CellOrder::RowMajor, sparse_schema.cell_order().unwrap());
-        assert_eq!(TileOrder::RowMajor, sparse_schema.tile_order().unwrap());
-        assert!(!sparse_schema.allows_duplicates().unwrap());
-    }
-}
