+++ tiledb/serde/src/filter/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(any(test, feature = "proptest-strategies"))]
+pub mod strategy;
