+++ tiledb/serde/src/lib.rs
@@ -4,6 +4,7 @@ extern crate tiledb_proc_macro;
 extern crate tiledb_sys as ffi;
 
 pub mod array;
+pub mod filter;
 
 #[cfg(any(test, feature = "proptest-strategies"))]
 pub mod strategy;
